# Perf: optimize function pow in module_gint