--- a/source/module_hamilt_lcao/module_gint/grid_meshball.cpp
+++ b/source/module_hamilt_lcao/module_gint/grid_meshball.cpp
@@ -1,5 +1,4 @@
 #include "grid_meshball.h"
-#include "gint_tools.h"  // for pow(double, int)
 #include "module_base/memory.h"
 
 Grid_MeshBall::Grid_MeshBall()
@@ -128,7 +127,7 @@ double Grid_MeshBall::deal_with_atom_spillage(const double *pos)
 					cell[ip] = i*this->bigcell_vec1[ip] +
 						j*this->bigcell_vec2[ip] +
 						k*this->bigcell_vec3[ip];
-					dx += pow(cell[ip] - pos[ip], 2);
+					dx += (cell[ip] - pos[ip]) * (cell[ip] - pos[ip]);
 				}
 				r2 = std::min(dx, r2);
 			}