### Background
When profiling the abacus program with vtune, I found that the `pow` function consumed a significant portion of the `cal_dpsir_ylm` function's runtime. `std::pow` is an inefficient function, especially when the exponent is a small integer, as direct multiplication can be dozens of times faster than calling `std::pow`. Thus, I overloaded the `pow` function for cases where the exponent is a small integer, which has accelerated the computation of `cal_gint_force`.
Before optimizing, the flame graph of `cal_dpsir_ylm`:
![image](https://github.com/user-attachments/assets/7d4dbbc0-1ea7-479c-bddd-23175aa0f1fe)
After optimizing, the flame graph of `cal_dpsir_ylm`(the "_INTERN..." part represents `pow` function):
![image](https://github.com/user-attachments/assets/84494475-e840-4ee9-9c01-919100e716ef)
It's clear that the time portion spent in `pow` function dropped significantly after the overloaded of `pow`.
Here is the performance comparison (the testing example is tests/performance/P103_Si128_lcao and the testing command is "OMP_NUM_THREADS=12 mpirun -n 1 abacus"):
|  | before optimization | after optimization |
| --- | --- | --- |
| cal_gint_force(CPU)| 16.81s | 14.06s |
| cal_gint_force(GPU) | 3.91s | 3.01s |